In order for MAXVNODES_MAX to be an "int" on powerpc and sparc, we must cast PAGE_SIZE to an "int".  (Powerpc and sparc, unlike the other architectures, define PAGE_SIZE as a "long".)
Submitted by:	Andreas Tobler


Former-commit-id: db22f27a2ccd8b16dfc6d5006e60ff5de244fd52

+++ sys/kern/vfs_subr.c
@@ -1272,7 +1272,7 @@ begin_define
 define|#
 directive|define
 name|MAXVNODES_MAX
-value|(512 * (1024 * 1024 * 1024 / PAGE_SIZE / 16))
+value|(512 * (1024 * 1024 * 1024 / (int)PAGE_SIZE / 16))
 end_define
 
 begin_endif
